refactor(widget): remove unused components, move components
