+++ apps/web/src/services/dashboards/shared/dashboard-widget-container/WidgetViewModeSidebar.vue
